refactor(): move classes and interface in the package entity and interface
