+++ app/src/main/java/com/rossel/android/sdk/mytictactoe/domain/entity/VerifierVerticalRow.kt
@@ -1,0 +1,11 @@
+package com.rossel.android.sdk.mytictactoe.domain.entity
+
+import com.rossel.android.sdk.mytictactoe.domain.enums.Player
+import com.rossel.android.sdk.mytictactoe.domain.interfaces.IVerifierRow
+
+class VerifierVerticalRow: IVerifierRow {
+    override fun verifierRow(board: List<Player>, player: Player): Boolean =
+        (board[0] == player && board[3] == player && board[6] == player)
+                || (board[1] == player && board[4] == player && board[7] == player)
+                || (board[2] == player && board[5] == player && board[8] == player)
+}
